Corrected training methods
Training was not working with the SubprocVecEnv, so it was replaced with the use of the make_vec_enc function used previously in my ROOTRL project.

I also fiddled around with training and may in the future add more pre-programmed agents to the training pool than just the 'base' model.