Rotary Emb / Attention Visuals
Attempted to implement a special type of positional embedding described by Monroe and Chalmers' "Mastering Chess with a Transformer Model" (2024), but Copilot and I could not quite come up with a quick, vector/einsum-based implementation. I will use a Rotary Positional Embedding instead, which seems to work well enough here.

Made a pretty long notebook attempting to visualize the attention weights and values of different layers in the model, as shown by Monroe and Chalmers' paper for chess.

- Added a completely random-choosing opponent for training and testing, and added it to the pool of default training opponents